refactor: add try/except/finally function for test_save_passwords_txt

+++ tests/test_function.py
@@ -72,26 +72,28 @@ def test_save_passwords_json():
 
 def test_save_passwords_txt():
     """Test save_passwords_txt function."""
-    test_passwords = [
-        {"service": "Google", "login": "user@gmail.com", "password": "P@ssw0rd!"},
-        {"service": "GitHub", "login": "username", "password": "Abc123!"}
-    ]
-
+    test_passwords = get_test_passwords()
     file_path = "text_passwords.txt"
 
-    save_passwords_txt(file_path, test_passwords), "File was not created!"
+    try:
+        save_passwords_txt(file_path, test_passwords), "File was not created!"
 
-    with open(file_path, "r", encoding="utf-8") as f:
-        lines = f.readlines()
-        assert len(lines) == 2, "Wrong number of lines!"
+        with open(file_path, "r", encoding="utf-8") as f:
+            lines = f.readlines()
+            assert len(lines) == 2, "Wrong number of lines!"
 
-        expected = "Service: Google | Login/email: user@gmail.com | Password: P@ssw0rd!\n"
-        assert lines[0] == expected, "Wrong content in first line!"
+            expected = "Service: Google | Login/email: user@gmail.com | Password: P@ssw0rd!\n"
+            assert lines[0] == expected, "Wrong content in first line!"
 
-    print("✅ All tests for txt format passed!")
+        print("✅ All tests for txt format passed!")
 
-    os.remove(file_path)
-    print("🗑️ Test file removed.")
+    except Exception as e:
+        print(f"❌ Test failed: {e}")
+        raise
+    finally:
+        if os.remove(file_path):
+            os.remove(file_path)
+            print("🗑️ Test file removed.")
 
 
 if __name__ == "__main__":
